microservices/user_manager/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.76.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
protobuf==6.33.1
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
psycopg2-binary
requests
python-dotenv